==> code/skill-matching/code/1_create_resume_corpus.csv.ipynb <==
import os,sys,re
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm
import codecs
import warnings
warnings.simplefilter('ignore')
import glob

#ファイルのロード
dir = "/workspace/data/input"
os.chdir("/workspace/data/input/resume_corpus/")
text_files = glob.glob("*.txt")
lab_files = glob.glob("*.lab")

text_data = list(np.zeros(np.array(text_files).shape).astype(str)) #.txtファイルのデータを入れるリスト
text_index_data = np.zeros(np.array(text_files).shape).astype(int) #ファイル番号を入れる配列
for i, text_file in enumerate(tqdm(text_files)): 
    index = int(os.path.splitext(os.path.basename(text_file))[0])
    with codecs.open(text_file, "r", "Shift-JIS", "ignore") as f:
        text = f.readline()
    text_index_data[i] = index
    text_data[i] = text

lab_data = list(np.zeros(np.array(lab_files).shape).astype(str)) #.labファイルのデータを入れるリスト
lab_index_data = np.zeros(np.array(lab_files).shape).astype(int) #ファイル番号を入れる配列
for i,lab_file in enumerate(tqdm(lab_files)):
    index = int(os.path.splitext(os.path.basename(lab_file))[0])
    with codecs.open(lab_file, "r", "Shift-JIS", "ignore") as f:
        text = f.readlines()
    lab_index_data[i] = index
    lab_data[i] = text 
    for j in range(len(lab_data[i])):
        lab_data[i][j] = lab_data[i][j].replace("\n","")
#indexとlabel要素をzipした辞書を作成
d = {k: v for k, v in zip(lab_index_data,lab_data)}

#ラベルにある職種にidを付与する
job2id = {}
ind = 0
for job in lab_data:
    for j in job:
        if(j not in job2id):
            job2id[j] = ind
            ind = ind + 1

#jobをindに変換
for id_key in d.keys():
    for i in range(len(d[id_key])):
        for (job_id,job) in zip(job2id.keys(),job2id.values()):
            d[id_key][i] = d[id_key][i].replace(job_id,str(job))
            
#labelをone-hot-encodingする
onehot_dict = {}
for i in range(0,10):
    onehot_dict[i] = 0
for key in d.keys():
    for job_id in d[key]:
        if(int(job_id) in job2id.values()): 
            onehot_dict[int(job_id)] = 1 #該当する職種に0を付与
    d[key] = onehot_dict
    onehot_dict = {}
    # for i in range(0,10):
    #     onehot_dict[i] = 0


#作成した辞書をデータフレームに格納
df = pd.DataFrame(d).T
txt_df = pd.DataFrame(text_data,index=text_index_data,columns=["cv"])
_df = pd.merge(df,txt_df,on=df.index)
for (x,y) in zip(job2id.values(),job2id.keys()):
    _df = _df.rename(columns={x:y})
_df.set_index("key_0",inplace=True)

_df.to_csv("/workspace/data/input/resume_corpus.csv")
